Split cloudformation class into two
