+++ Watchman.Engine/Generation/Generic/ICloudformationStackDeployer.cs
@@ -1,0 +1,9 @@
+﻿using System.Threading.Tasks;
+
+namespace Watchman.Engine.Generation.Generic
+{
+    public interface ICloudformationStackDeployer
+    {
+        Task DeployStack(string name, string body, bool isDryRun);
+    }
+}
